refactor: study list filter, query fix (#97)
