+++ src/components/CommunityPost/List/SearchSheet/SearchSheet.tsx
@@ -9,7 +9,7 @@ import Tag from "@/components/common/Tag/Tag";
 import { getLocalStorage, setLocalStorage } from "@/utils/storage";
 
 import type { CommonSheetProps } from "@/types/common";
-import type { UpdateFilterSelectedTypeHandlerType } from "@/types/common";
+import type { UpdateFilterSelectedTypeHandlerType } from "@/types/filter";
 
 interface SearchSheetProps extends CommonSheetProps {
   searchValue: string;
