+++ src/hooks/api/study/useStudyListQuery.ts
@@ -6,10 +6,10 @@ import { GET, createInit } from "@/apis/httpMethod";
 import { END_POINTS } from "@/constants/api";
 import { QUERY_KEYS } from "@/constants/queryKey";
 
-import type { StudyListFilterType } from "@/types/study";
+import type { FilterSelectedType } from "@/types/filter";
 import type { GetStudyListResponseType } from "@/types/study";
 
-const getStudyList = async (param?: StudyListFilterType) => {
+const getStudyList = async (param?: FilterSelectedType) => {
   const queryString = new URLSearchParams(Object.entries(param ?? {}));
 
   const data = await GET<GetStudyListResponseType>(
@@ -19,13 +19,14 @@ const getStudyList = async (param?: StudyListFilterType) => {
 
   return data;
 };
+
 export const studyListQueryOptions = (
-  param?: StudyListFilterType,
+  param?: FilterSelectedType,
 ): UseSuspenseQueryOptions<GetStudyListResponseType> => ({
   queryKey: [QUERY_KEYS.STUDY_LIST],
   queryFn: () => getStudyList(param),
 });
 
-export function useStudyListQuery(param?: StudyListFilterType) {
+export function useStudyListQuery(param?: FilterSelectedType) {
   return useSuspenseQuery(studyListQueryOptions(param));
 }
